+++ CHANGES.md
@@ -1,6 +1,7 @@
 v0.1.7 (dev)
 ============
 + add wiki page and more information about decomposing variants
++ clarify warning on non ACGT (#23)
 
 v0.1.6
 ======